Fix for bug #118
According to the debugger the following sequence would happen.

1. A packet is received.
2. we start handling the packet.
3. The packet is reset.
4. we continue handling the packet.
5. The application crashes due to a seg fault.

However this sequence should never happen because there is only one thread at work
and that thread must finish handling the packet before reseting.
Basically line #2 is being executed before line #1 which should never happen.
I don't know why this is happening, my guess is a bug in the compiler or a bug in the MFC thread library.

I rearranged the code a bit and the problem no longer occurs.

Closes #118