Fix for bug #133
The mediator would not exit the game when the application is being closed.
This would result in a bunch of threads trying to send data over a closed socket.
Which would in turn result in a infinite loop.

I fixed this by closing the game before disconnecting the sockets.
I also by added a check to the send functions,
the send will fail if we are in the process of disconnecting.

Closes #133