Added a homegrown unit test framework and a sample test suite.
The framework is limited at the moment but I can improve on it as needed.
The easiest way to run the tests is to run the following command while in the Debug directory.

./Chat.exe --test; cat TestResults.txt

The following output should appear:

Sample Test Suite
3 out of 6 tests failed.
testAssertEquals : SUCCESS
testAssertEqualsFail : FAILED - 1 is not equal to 2
testAssertTrue : SUCCESS
testAssertTrueFail : FAILED - False is not true
testAssertFalse : SUCCESS
testAssertFalseFail : FAILED - True is not false