Fixed not being able to reconnect to peer.
This solution is more of a work around. The network is using the "DoModal"
thread to accept a connection and is trying to display a message in the
chat box which requires the "DoModal" thread. This results in a deadlock. By
reseting the peer's username when we disconnect we are no longer trying to
display "Oldname has changed his name to Newname" when acceptiong a connection.
This solved the deadlock. The changed name message shouldn't appear anyway
because the "Player has connected" message will be displayed.

Closes #249